backend/requirements.txt
fastapi
uvicorn
langgraph
langchain
langchain-groq
groq
pyannote.audio
pydantic
python-dotenv
python-multipart
soundfile
